tsmf: hide the plugin struct and use freerdp types.
